Use ShopUser prefix because validators can't be use with AdminUser
