+++ src/Sylius/Bundle/ApiBundle/Resources/config/services/validator.xml
@@ -115,21 +115,21 @@
             <tag name="validator.constraint_validator" alias="sylius_api_shipment_already_shipped" />
         </service>
 
-        <service id="sylius.api.validator.shop.user_exists" class="Sylius\Bundle\ApiBundle\Validator\Constraints\UserExistsValidator">
+        <service id="Sylius\Bundle\ApiBundle\Validator\Constraints\ShopUserExistsValidator">
             <argument type="service" id="sylius.canonicalizer" />
             <argument type="service" id="sylius.repository.shop_user" />
-            <tag name="validator.constraint_validator" alias="sylius_api_user_exists" />
+            <tag name="validator.constraint_validator" alias="sylius_api_shop_user_exists" />
         </service>
 
-        <service id="sylius.api.validator.shop.reset_password_token_exists" class="Sylius\Bundle\ApiBundle\Validator\Constraints\ResetPasswordTokenExistsValidator">
+        <service id="Sylius\Bundle\ApiBundle\Validator\Constraints\ShopUserResetPasswordTokenExistsValidator">
             <argument type="service" id="sylius.repository.shop_user" />
-            <tag name="validator.constraint_validator" alias="sylius_api_reset_password_token_exists" />
+            <tag name="validator.constraint_validator" alias="sylius_api_shop_user_reset_password_token_exists" />
         </service>
 
-        <service id="sylius.api.validator.shop.reset_password_token_not_expired" class="Sylius\Bundle\ApiBundle\Validator\Constraints\ResetPasswordTokenNotExpiredValidator">
+        <service id="Sylius\Bundle\ApiBundle\Validator\Constraints\ShopUserResetPasswordTokenNotExpiredValidator">
             <argument type="service" id="sylius.repository.shop_user" />
             <argument type="string">%sylius.shop_user.token.password_reset.ttl%</argument>
-            <tag name="validator.constraint_validator" alias="sylius_api_reset_password_token_not_expired" />
+            <tag name="validator.constraint_validator" alias="sylius_api_shop_user_reset_password_token_not_expired" />
         </service>
 
         <service id="Sylius\Bundle\ApiBundle\Validator\Constraints\ShopUserNotVerifiedValidator">
