+++ src/Sylius/Bundle/ApiBundle/Resources/config/validation/AccountResetPassword.xml
@@ -49,10 +49,10 @@
             <constraint name="NotBlank">
                 <option name="groups">sylius</option>
             </constraint>
-            <constraint name="Sylius\Bundle\ApiBundle\Validator\Constraints\ResetPasswordTokenExists">
+            <constraint name="Sylius\Bundle\ApiBundle\Validator\Constraints\ShopUserResetPasswordTokenExists">
                 <option name="groups">sylius</option>
             </constraint>
-            <constraint name="Sylius\Bundle\ApiBundle\Validator\Constraints\ResetPasswordTokenNotExpired">
+            <constraint name="Sylius\Bundle\ApiBundle\Validator\Constraints\ShopUserResetPasswordTokenNotExpired">
                 <option name="groups">sylius</option>
             </constraint>
         </property>
