+++ src/Sylius/Bundle/ApiBundle/Resources/config/validation/RequestResetPasswordToken.xml
@@ -25,7 +25,7 @@
             </constraint>
         </property>
         <property name="email">
-            <constraint name="Sylius\Bundle\ApiBundle\Validator\Constraints\UserExists">
+            <constraint name="Sylius\Bundle\ApiBundle\Validator\Constraints\ShopUserExists">
                 <option name="groups">
                     <value>sylius</value>
                 </option>
