+++ src/Sylius/Bundle/ApiBundle/Resources/translations/validators.en.yaml
@@ -33,5 +33,5 @@ sylius:
         not_found: 'The shipping method with %code% code does not exist.'
         not_available: 'The shipping method %name% is not available for this order. Please reselect your shipping method.'
         shipping_address_not_found: 'Order should be addressed first.'
-    user:
+    shop_user:
         not_found: 'The user with %email% email does not exist.'
