+++ src/Sylius/Bundle/ApiBundle/spec/Validator/Constraints/ShopUserExistsValidatorSpec.php
@@ -14,18 +14,21 @@ declare(strict_types=1);
 namespace spec\Sylius\Bundle\ApiBundle\Validator\Constraints;
 
 use PhpSpec\ObjectBehavior;
-use Sylius\Bundle\ApiBundle\Validator\Constraints\ResetPasswordTokenExists;
+use Sylius\Bundle\ApiBundle\Validator\Constraints\ShopUserExists;
+use Sylius\Component\User\Canonicalizer\CanonicalizerInterface;
 use Sylius\Component\User\Model\UserInterface;
 use Sylius\Component\User\Repository\UserRepositoryInterface;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidatorInterface;
 use Symfony\Component\Validator\Context\ExecutionContextInterface;
 
-final class ResetPasswordTokenExistsValidatorSpec extends ObjectBehavior
+final class ShopUserExistsValidatorSpec extends ObjectBehavior
 {
-    function let(UserRepositoryInterface $userRepository): void
-    {
-        $this->beConstructedWith($userRepository);
+    function let(
+        CanonicalizerInterface $canonicalizer,
+        UserRepositoryInterface $userRepository
+    ): void {
+        $this->beConstructedWith($canonicalizer, $userRepository);
     }
 
     function it_is_a_constraint_validator(): void
@@ -42,7 +45,7 @@ final class ResetPasswordTokenExistsValidatorSpec extends ObjectBehavior
         ;
     }
 
-    function it_throws_an_exception_if_constraint_is_not_a_resetPasswordTokenExists_constraint(): void
+    function it_throws_an_exception_if_constraint_is_not_a_userNotFound_constraint(): void
     {
         $this
             ->shouldThrow(\InvalidArgumentException::class)
@@ -52,37 +55,43 @@ final class ResetPasswordTokenExistsValidatorSpec extends ObjectBehavior
     }
 
     function it_does_not_add_violation_if_user_exist(
+        CanonicalizerInterface $canonicalizer,
         UserRepositoryInterface $userRepository,
         ExecutionContextInterface $executionContext,
         UserInterface $user
     ): void {
         $this->initialize($executionContext);
 
-        $value = 'token';
+        $value = 'sylius@example.com';
 
-        $userRepository->findOneBy(['passwordResetToken' => 'token'])->willReturn($user);
+        $canonicalizer->canonicalize('sylius@example.com')->willReturn('sylius@example.com');
+
+        $userRepository->findOneByEmail('sylius@example.com')->willReturn($user);
 
         $executionContext
-            ->addViolation('sylius.reset_password.invalid_token', ['%token%' => 'token'])
+            ->addViolation('sylius.user.not_found', ['%email%' => 'sylius@example.com'])
             ->shouldNotBeCalled();
 
-        $this->validate($value, new ResetPasswordTokenExists());
+        $this->validate($value, new ShopUserExists());
     }
 
-    function it_adds_violation_if_reset_password_token_does_not_exist(
+    function it_adds_violation_if_user_does_not_exist(
+        CanonicalizerInterface $canonicalizer,
         UserRepositoryInterface $userRepository,
         ExecutionContextInterface $executionContext
     ): void {
         $this->initialize($executionContext);
 
-        $value = 'token';
+        $value = 'sylius@example.com';
 
-        $userRepository->findOneBy(['passwordResetToken' => 'token'])->willReturn(null);
+        $canonicalizer->canonicalize('sylius@example.com')->willReturn('sylius@example.com');
+
+        $userRepository->findOneByEmail('sylius@example.com')->willReturn(null);
 
         $executionContext
-            ->addViolation('sylius.reset_password.invalid_token', ['%token%' => 'token'])
+            ->addViolation('sylius.shop_user.not_found', ['%email%' => 'sylius@example.com'])
             ->shouldBeCalled();
 
-        $this->validate($value, new ResetPasswordTokenExists());
+        $this->validate($value, new ShopUserExists());
     }
 }
