+++ src/Sylius/Bundle/ApiBundle/spec/Validator/Constraints/ShopUserResetPasswordTokenExistsValidatorSpec.php
@@ -14,21 +14,18 @@ declare(strict_types=1);
 namespace spec\Sylius\Bundle\ApiBundle\Validator\Constraints;
 
 use PhpSpec\ObjectBehavior;
-use Sylius\Bundle\ApiBundle\Validator\Constraints\UserExists;
-use Sylius\Component\User\Canonicalizer\CanonicalizerInterface;
+use Sylius\Bundle\ApiBundle\Validator\Constraints\ShopUserResetPasswordTokenExists;
 use Sylius\Component\User\Model\UserInterface;
 use Sylius\Component\User\Repository\UserRepositoryInterface;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidatorInterface;
 use Symfony\Component\Validator\Context\ExecutionContextInterface;
 
-final class UserExistsValidatorSpec extends ObjectBehavior
+final class ShopUserResetPasswordTokenExistsValidatorSpec extends ObjectBehavior
 {
-    function let(
-        CanonicalizerInterface $canonicalizer,
-        UserRepositoryInterface $userRepository
-    ): void {
-        $this->beConstructedWith($canonicalizer, $userRepository);
+    function let(UserRepositoryInterface $userRepository): void
+    {
+        $this->beConstructedWith($userRepository);
     }
 
     function it_is_a_constraint_validator(): void
@@ -45,7 +42,7 @@ final class UserExistsValidatorSpec extends ObjectBehavior
         ;
     }
 
-    function it_throws_an_exception_if_constraint_is_not_a_userNotFound_constraint(): void
+    function it_throws_an_exception_if_constraint_is_not_a_resetPasswordTokenExists_constraint(): void
     {
         $this
             ->shouldThrow(\InvalidArgumentException::class)
@@ -55,43 +52,37 @@ final class UserExistsValidatorSpec extends ObjectBehavior
     }
 
     function it_does_not_add_violation_if_user_exist(
-        CanonicalizerInterface $canonicalizer,
         UserRepositoryInterface $userRepository,
         ExecutionContextInterface $executionContext,
         UserInterface $user
     ): void {
         $this->initialize($executionContext);
 
-        $value = 'sylius@example.com';
+        $value = 'token';
 
-        $canonicalizer->canonicalize('sylius@example.com')->willReturn('sylius@example.com');
-
-        $userRepository->findOneByEmail('sylius@example.com')->willReturn($user);
+        $userRepository->findOneBy(['passwordResetToken' => 'token'])->willReturn($user);
 
         $executionContext
-            ->addViolation('sylius.user.not_found', ['%email%' => 'sylius@example.com'])
+            ->addViolation('sylius.reset_password.invalid_token', ['%token%' => 'token'])
             ->shouldNotBeCalled();
 
-        $this->validate($value, new UserExists());
+        $this->validate($value, new ShopUserResetPasswordTokenExists());
     }
 
-    function it_adds_violation_if_user_does_not_exist(
-        CanonicalizerInterface $canonicalizer,
+    function it_adds_violation_if_reset_password_token_does_not_exist(
         UserRepositoryInterface $userRepository,
         ExecutionContextInterface $executionContext
     ): void {
         $this->initialize($executionContext);
 
-        $value = 'sylius@example.com';
+        $value = 'token';
 
-        $canonicalizer->canonicalize('sylius@example.com')->willReturn('sylius@example.com');
-
-        $userRepository->findOneByEmail('sylius@example.com')->willReturn(null);
+        $userRepository->findOneBy(['passwordResetToken' => 'token'])->willReturn(null);
 
         $executionContext
-            ->addViolation('sylius.user.not_found', ['%email%' => 'sylius@example.com'])
+            ->addViolation('sylius.reset_password.invalid_token', ['%token%' => 'token'])
             ->shouldBeCalled();
 
-        $this->validate($value, new UserExists());
+        $this->validate($value, new ShopUserResetPasswordTokenExists());
     }
 }
